+++ src/comp.h
@@ -59,7 +59,7 @@ void copybest();
 bool init(optionStruct * options);
 void brown();
 void printparms(int Nnodes,double weight, optionStruct * options);
-void betterfound(int Nnodes,double weight,int swath,int iterations,int blobs,int lines,double fraction,int fraclines,bool improvement,optionStruct * options);
+void betterfound(int Nnodes,double weight,int swath,int iterations,int blobs,size_t lines,double fraction,size_t fraclines,bool improvement,optionStruct * options);
 void worsefound();
 
 extern int (*comp)(const vertex * a,const vertex * b);
